week-14/pyproject.toml
[project]
name = "week-14"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser (>=6.0.11,<7.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.65.3,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
